refactor: move PieChartColor to model package
