+++ app/src/main/java/otus/homework/customview/presentation/model/PieChartAngle.kt
@@ -1,7 +1,5 @@
 package otus.homework.customview.presentation.model
 
-import otus.homework.customview.presentation.piechartview.PieChartColor
-
 data class PieChartAngle(
     val id: Int,
     val category: String,
